requirements.txt
streamlit
streamlit-option-menu
cohere
python-docx